add : section3 및 회원가입, 카카오 로그인,회원가입 완료
