+++ .env
@@ -1,0 +1,3 @@
+REACT_APP_KAKAOLOGIN_API_KEY = 86483f30e78c6016d89913f11cd358ce
+REACT_APP_KAKAOLOGIN_REDIRECT_URI = http://localhost:3000/KakaoLogin
+
