+++ package.json
@@ -12,6 +12,7 @@
     "@types/react-dom": "^18.2.1",
     "axios": "^1.4.0",
     "react": "^18.2.0",
+    "react-cookie": "^4.1.1",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.11.0",
     "react-scripts": "5.0.1",
