+++ src/App.tsx
@@ -1,10 +1,15 @@
 import React from "react";
+import { CookiesProvider } from "react-cookie";
 import { RouterProvider } from "react-router-dom";
 import "./App.css";
 import Router from "./Router";
 
 function App() {
-  return <RouterProvider router={Router} />
+  return (
+    <CookiesProvider>
+      <RouterProvider router={Router} />
+    </CookiesProvider>
+  );
 }
 
 export default App;
