+++ src/styles/Main/mainSection03.module.css
@@ -81,7 +81,8 @@
   /* justify-content: center; */
   align-items: center;
   width: 480px;
-  min-height: 800px;
+  min-height: 650px;
+  padding: 10px 0;
   background-color: #fff;
   border-radius: 30px;
 }
@@ -110,7 +111,7 @@
 }
 
 .signUpInput {
-    width: 100%;
+  width: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -148,4 +149,27 @@
   color: red;
 }
 
+.signUpInput button {
 
+  padding: 8px 80px;
+  font-size: 20px;
+  font-weight: bold;
+  border-radius: 30px;
+  color: #fff;
+  background-color: #414141;
+}
+
+.kakaoLogin {
+  margin-top: 20px;
+}
+.kakaoLogin button {
+  width: 183px;
+  height: 45px;
+  background-color: #fff;
+  cursor: pointer;
+}
+.kakaoLogin img {
+  width: 100%;
+  height: 100%;
+  border-radius: 30px;
+}
